First attempt at remolding tutorial in interactive Jupyter notebooks
Created a tutorial notebook for the QueryBuilder, loosely based on the
existing tutorial. The idea is that with a notebook, we can hand the
user a single resource that will give them all the information that they
need, offer examples, but also the opportunity to try and use the code
themselves by offering them exercises. That is also why there are two
versions of the notebook. The version in the 'full' directory, includes
all the soltuions to the proposed exercises. Since handing this one to
the user would defeat the purpose of asking them to complete the exercise,
there is another version in 'user' which has parts of the solution code
removed or obscured.

Obviously, correctly maintaining these two version in parallel can be a
headache, but I am not yet aware of a better and easier solution.

Also the current notebook is perhaps a little light on exercises. We may
also want to come up with some kind of method of providing optional hints
or even solutions that are initially hidden in some way. It would be best
if all of this could be done in the same notebook, so that it really
remains a stand-alone learning resource.