Add a new tutorial notebook to calculate bandstructure with pw.x through workchains
The notebook is designed to be a pure demonstration. It requires minimal imports
and setup and simply uses COD importer to import several simple structures and
launches the PwBandsWorkChain that will try and calculate the band structure
using QE's pw. Note that the underlying PwBaseWorkChain is a bare essential
workchain and does not implement any sophisticated restarting mechanism.
Therefore only structures that will successfully converge in a single PwCalculation
will be computed successfully.