Add a band structure and basic QE pw.x workchain
These workchains serve as a demonstration, to be run from a jupyter
notebook, which can also be found in this repo, of how to use AiiDA
workchains to calculate a bandstructure. Note that the underlying
PwBaseWorkChain is very primitive and does not handle submission or
calculation failures. It does however provide the scaffolding for this
functionality to be easily implemented