Change the pw codename string to match that of the tutorial AWS
In this way, for the tutorial machine, the notebooks won't have
to be adapted. Of course if you want to run these on your local
machine you will have to make sure to correct them to match your
local AiiDA database. The same goes for the PwBandsWorkChain and
PwBaseWorkChain in workchain/quantumespresso/pw which also contain
hardcoded import, codename and SSSP string values