Update to the band notebook ()
Finished the conversion to aiida 1.0b3, and the latest quantumespresso plugin.
The workchain now runs correctly with the SSSP v1.1, efficiency.
Tests by me and marco.

Minor issue: for one example (GaAs) the default time and computing power
is not enough to finish the job, so the obtained band structure is
incomplete (path L-W-X is empty).
That is the longest example we provide (it takes almost 2h on my workstation)
An increase of the default from 1800s to 3600s solves the problem for me.
This could be implemented in 3 possible ways:
- be modified in the qe plugin
- be provided as an option in the submission, at the price of a more complicated
input for the user of the notebook, but the advantage of showing that modification
to the turn-key solution are allowed.

results_GaAs = launch.run(
    PwBandStructureWorkChain,
    code=code,
    structure=structure_GaAs,
    scf_options = Dict(dict={
        'resources': {
            'num_machines': 1
        },
        'max_wallclock_seconds' : 3600
    })
)

- provide a modification of the protocol with reduced accuracy ('fast' keyword)
for this specific case. This also complicates the input, but provided additional info
about the plugin